+++ misc/dev-server.py
@@ -1,0 +1,8 @@
+"""
+Just a script to run a server in tests/server/. :)
+"""
+import os
+
+cmd = f"cd tests/server;python3 -m http.server"
+
+os.system(cmd)